+++ src/cpu/o3/fetch.hh
@@ -414,7 +414,7 @@ class DefaultFetch
     LVPredUnit *loadPred;
 
     // For Trace-Based Implementation, need to track trace ID across cycles
-    unsigned int currentTraceID = 0;
+    uint64_t currentTraceID = 0;
 
   private:
     bool usingTrace = false;
